refactored js to modules
