+++ src/index.html
@@ -17,6 +17,7 @@
       defer
       type="text/javascript"
       src="scripts/main.js"
+
     ></script>
   </head>
   <body>
@@ -202,12 +203,7 @@
 
             <button class="slider-about__btn slider-about__btn--left">&larr;</button>
             <button class="slider-about__btn slider-about__btn--right">&rarr;</button>
-            <div class="dots">
-              <!-- <div class="dots__dot"></div>
-              <div class="dots__dot"></div>
-              <div class="dots__dot"></div>
-              <div class="dots__dot"></div> -->
-            </div>
+            <div class="dots"></div>
           </div>
         </div>
       </section>
