+++ src/scripts/modules/checkMedia.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const checkMedia = function(callback) {
+  const mediaQuery = window.matchMedia('(max-width: 666px)');
+
+  if (mediaQuery.matches) {
+    callback; // eslint-disable-line
+  }
+};
+
+module.exports = { checkMedia };
