+++ src/scripts/modules/removeScroll.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const removeScroll = function(hash) {
+  window.addEventListener('hashchange', () => {
+    if (window.location.hash === hash) {
+      document.body.classList.add('page__body--with-menu');
+    } else {
+      document.body.classList.remove('page__body--with-menu');
+    }
+  });
+};
+
+module.exports = { removeScroll };
